+++ src/index.ts
@@ -1111,6 +1111,7 @@ function parameter(name: string, schema: z.ZodType, symbol = name) {
       mode: 'strict' as const,
       typeSymbol: `dsh-airdrop#${symbol}`,
       schema,
+      create: () => schema,
     },
   }
 }
@@ -1127,6 +1128,7 @@ function invocation(method: string, parameters: readonly ReturnType<typeof param
       mode: 'strict' as const,
       typeSymbol: `dsh-airdrop#${resultSymbol}`,
       schema,
+      create: () => schema,
     },
   }
 }